+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Volumes now provide `get_journal_title()` to fetch the journal title from the venue metadata if it's not explicit set.
 - Papers now have attributes `bibtype` and `web_url`.
 
+### Changed
+
+- Bugfix: `parse_id()` now parses old-style frontmatter IDs correctly.
+
 ## [0.4.2] — 2023-10-21
 
 ### Added